+++ dino_catalog/dinosaur_row_reader.rb
@@ -1,7 +1,7 @@
 class DinosaurRowReader
   attr_accessor :values
 
-  def parse_row(row)
+  def parse_row(_row)
     raise NotImplementedError, "this base class should never be initiated"
   end
 end
